refactored the forgetpassword and touched the repository file
