+++ src/modules/auth/model.js
@@ -1,4 +1,4 @@
-const {getUser, updateUser, createUser } = require('./repository')
+const {getUser, updateUser, createnewUser } = require('./repository')
 const { hashPassword, comparePasswords } = require('../../utils/hasher')
 const { generateToken, generateEmailVerificationLink, generatePasswordResetLink, verifyLink } = require('../../utils/token')
 const  MailService  = require('./mailService')
@@ -6,7 +6,7 @@ const  MailService  = require('./mailService')
 module.exports = class Auth {
     static async createUser (email, password) {
             // check if user exists
-            let userExists = await Auth.getUser(email)
+            let userExists = await Auth.checkUser(email)
             if (userExists){
                 throw new Error("User already exists")
             }
@@ -15,7 +15,7 @@ module.exports = class Auth {
             let hashedPassword = await hashPassword(password)
             
             //save details to database
-            let user = await createUser(email, hashedPassword)
+            let user = await createnewUser(email, hashedPassword)
         
             //generate verification email
             let verificationLink = await generateEmailVerificationLink(user)
@@ -28,9 +28,29 @@ module.exports = class Auth {
     }
 
     
-    static async getUser(email) {
+    static async checkUser(email) {
         return await getUser(email)
     }
+    
+    static async getUser(email) {
+        let user = await getUser(email)
+
+        //filter user object
+        user = {
+            id: user._id,
+            email: user.email
+        }
+
+        //generate password reset email
+        let passwordResetLink = await generatePasswordResetLink(user)
+        let mailData = {
+            to: user.email,
+            passwordResetLink
+        }
+        await MailService.sendPasswordResetMail(mailData)
+
+    }
+    
 
     //query is the search parameter, data is the details to be updadted
     static async updateUser(filter, update){
